requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
tensorflow>=2.15.0
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0
google-generativeai>=0.4.0
fpdf>=1.7.2
python-dotenv>=1.0.1